// src/types/Enums/Descriptions/Tower/AlchemistsDescription.ts
export enum AlchemistsDescription {
  FAFNER = 'Говорят, что Фафнер превратился в джинна лет тысячу, а то и больше назад, когда, наткнувшись на бутылку с джинном, потребовал от её обитателя сделать его таким же могущественным, как он сам. Это желание было исполнено, и с тех пор Фафнер верой и правдой служит правителям Бракады.',
  IONA = 'Иона давно доказала своё превосходство над многими из её собратьев, но из-за лёгкого и весёлого характера её больше любят, чем боятся.',
  JOSEPHINE = 'Жозефине первой удалось успешно оживить каменного голема. Для этого она разработала специальный процесс, который оказался гораздо сложнее технологии оживления горгулий, взятой ею за основу.',
  NEELA = 'Нила, пожалуй, сильнейшая среди себе подобных и может без особого вреда переносить невероятные повреждения. Ей каким-то образом удаётся передавать эту способность и своим войскам.',
  PIQUEDRAM = 'Большую часть своих занятий алхимией Пикедрам посвятил изучению разных малоценных камней, и потому неудивительно, что на службе Бракаде ему удалось заметно улучшить начальную конструкцию горгулий.',
  RISSA = 'Рисса первой из алхимиков овладела искусством превращения бесполезных мягких металлов в ртуть. Хотя сейчас она командует собственной армией, в свободное время она продолжает от души экспериментировать с различными субстанциями.',
  THANE = 'Никто доподлинно не знает, кто такой Тан и откуда он родом. Он преподаёт алхимию в Бракаде уже столько лет, что даже старейшины не помнят времён, когда его тут не было.',
  TOROSAR = 'Будучи алхимиком по образованию, Торосар, тем не менее, всегда живо интересовался тактикой сражений и осад. Его чаще можно было застать за трактатами по этим вопросам, чем за фолиантами по алхимии.'
}